Split menus into Windows and Linux
