+++ LinuxCmdMenus.py
@@ -1,0 +1,10 @@
+import curses
+from cur
+
+
+
+stdscr = curses.initscr()
+
+#row, column, string
+def print_at(r, c, s):
+    stdscr.
